backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.4.0
pypdf==4.3.1
pdf2image==1.17.0
Pillow==10.4.0
reportlab==4.2.2
